Fix lesson-1-4

+++ lesson-1-4.cpp
@@ -2,14 +2,14 @@
 #include <cmath>
 
 int main() {
-    int r;
+    float r;
 
     std::cout << "Enter r:";
     std::cin >> r;
 
-    int d = 2 * r;
-    int i = 2 * M_PI * r;
-    int s = M_PI * pow(r, 2);
+    float d = 2 * r;
+    float i = 2 * M_PI * r;
+    float s = M_PI * pow(r, 2);
 
     std::cout << "Circle diameter = " << d << std::endl;
     std::cout << "Circumference = " << i << std::endl;
